initial work of moving code from ZE to ZE API for VSC creation of team config file
